Updated for backwards compatibility
Within createGiottoImage:
Added do_manual flag argument
If TRUE, only minmax values will be created and adj values given as arguments will be used
If FALSE and either spatlocs or gobject is given, adj values will be auto generated
If FALSE and no spatlocs or gobject is given, adj values will default to those given as arguments or default values.

Added do_manual_adj argument to createGiottoVisiumObject so that it can be passed through to createGiottoImage if manual alignment is desired.