giotto points update
Updated gpoints object creation behavior so that the first 'character' column seen is taken as the feat_ID column. The first two 'numeric' columns are taken as the x and y columns respectively. If any columns are given as 'feat_ID', 'x', or 'y', this behavior is skipped and those columns are directly taken.