objects and accessors update
Added all s4 objects other than spatial_enrichment.
Updated references to the spatial_locs slot within the code to use the accessor instead.

Still need to make sure that code that edits the spatial locations is returning as the S4 instead of the S3, which can lose the provenance info